refactor: create turn package with game turn and role actions
