+++ common/src/main/kotlin/fr/agrouagrou/common/turn/FortuneTellerActions.kt
@@ -1,5 +1,6 @@
-package fr.agrouagrou.common.state
+package fr.agrouagrou.common.turn
 
+import fr.agrouagrou.common.GameState
 import fr.agrouagrou.common.player.PlayerManager
 import fr.agrouagrou.common.player.Role
 import kotlinx.coroutines.flow.MutableStateFlow
