+++ common/src/main/kotlin/fr/agrouagrou/common/turn/GameTurn.kt
@@ -1,10 +1,7 @@
-package fr.agrouagrou.common
+package fr.agrouagrou.common.turn
 
+import fr.agrouagrou.common.GameState
 import fr.agrouagrou.common.player.PlayerManager
-import fr.agrouagrou.common.state.FortuneTellerActions
-import fr.agrouagrou.common.state.GameState
-import fr.agrouagrou.common.state.WerewolfActions
-import fr.agrouagrou.common.state.WitchActions
 import kotlinx.coroutines.flow.MutableStateFlow
 
 class GameTurn(val turnNumber: Int, gameState: MutableStateFlow<GameState>, playerManager: PlayerManager) {
