+++ common/src/main/kotlin/fr/agrouagrou/common/turn/RoleActions.kt
@@ -1,5 +1,6 @@
-package fr.agrouagrou.common.state
+package fr.agrouagrou.common.turn
 
+import fr.agrouagrou.common.GameState
 import fr.agrouagrou.common.player.Player
 import fr.agrouagrou.common.player.Role
 import kotlin.reflect.KClass
